refactor included readfish code to avoid confusion with installed module
